backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_cors
werkzeug
requests
razorpay
google-genai
python-dotenv
flask-socketio
eventlet